LayoutTests/editing/pasteboard/paste-blockquote-3.html

<style>
blockquote {
    color: blue;
    border-left: 2px solid blue;
    margin: 0px;
    padding: 0 0 0 20px;
}
</style>

<p>This tests pasting a blockquote into the middle of a paragraph.</p>
<div id="div" contenteditable="true">This should not be blockquoted. This should not be blockquoted.</div>

<script>
var sel = window.getSelection();
var div = document.getElementById("div");

sel.setPosition(div, 0);
sel.modify("move", "forward", "word");
sel.modify("move", "forward", "word");
sel.modify("move", "forward", "word");
sel.modify("move", "forward", "word");
sel.modify("move", "forward", "word");
sel.modify("move", "forward", "character");
document.execCommand("InsertHTML", false, "<blockquote type='cite'>&nbsp;This should not be blockquoted.<br>This should be blockquoted.<br>This should not be blockquoted.</blockquote>");
</script>
